[catalog] Fix mz_sources, progress type, subsource cluster_id, and builtin fingerprint
- parse_catalog_create_sql: detect progress subsources and extract
  of_source_id; use CROSS JOIN LATERAL to call it once per row
- Inline builtin sources as VALUES in mz_sources SQL so adding a
  builtin source changes the fingerprint and forces a migration
